Separate core logic as a package
