+++ PetAdoptions/payforadoption-go/go.mod
@@ -1,11 +1,11 @@
-module github.com/aws-samples/one-observability-demo/PetAdoptions/payforadoption-go
+module petadoptions
 
-go 1.14
+go 1.15
 
 require (
 	github.com/denisenkom/go-mssqldb v0.9.0
 	github.com/dghubble/sling v1.3.0
 	github.com/go-kit/kit v0.10.0
 	github.com/gofrs/uuid v3.3.0+incompatible
-	github.com/gorilla/mux v1.8.0
+	github.com/gorilla/mux v1.7.3
 )
