+++ PetAdoptions/payforadoption-go/payforadoption/domain.go
@@ -1,4 +1,4 @@
-package main
+package payforadoption
 
 import "time"
