+++ PetAdoptions/payforadoption-go/payforadoption/endpoint.go
@@ -1,4 +1,4 @@
-package main
+package payforadoption
 
 import (
 	"context"
