+++ PetAdoptions/payforadoption-go/payforadoption/service.go
@@ -1,4 +1,4 @@
-package main
+package payforadoption
 
 import (
 	"context"
