+++ PetAdoptions/payforadoption-go/payforadoption/transport.go
@@ -1,4 +1,4 @@
-package main
+package payforadoption
 
 import (
 	"context"
